+++ host/sgx/sgxquote.h
@@ -82,7 +82,7 @@ oe_result_t oe_tdx_verify_quote(
     const uint8_t* p_quote,
     uint32_t quote_size,
     const uint8_t* p_endorsements,
-    size_t endorsements_size,
+    uint32_t endorsements_size,
     time_t expiration_check_date,
     uint32_t* p_collateral_expiration_status,
     uint32_t* p_quote_verification_result,
@@ -90,5 +90,14 @@ oe_result_t oe_tdx_verify_quote(
     uint32_t qve_report_info_size,
     void* p_supplemental_data,
     uint32_t supplemental_data_size);
+
+oe_result_t oe_get_tdx_quote_verification_collateral(
+    const uint8_t* p_quote,
+    uint32_t quote_size,
+    uint8_t** pp_quote_collateral,
+    uint32_t* p_collateral_size);
+
+oe_result_t oe_free_tdx_quote_verification_collateral(
+    uint8_t* p_quote_collateral);
 
 #endif // _OE_SGXQUOTE_H