+++ CHANGELOG.md
@@ -44,6 +44,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - Note that the `oe_verify_evidence` with the plugin currently
     does not support input endorsements (must be `NULL`)
 
+- Added two APIs `oe_get_tdx_endorsements` and `oe_free_tdx_endorsements` to fetch
+  and free the endorsements for the given TDX quote. The APIs are for users who want
+  to manage (e.g., caching) the endorsements by themselves instead of relying on the
+  existing caching mechanisms (e.g., supported by DCAP).
+  See `openenclave/attestation/tdx/evidence.h` for more detail of the APIs.
+
 ## Changed
 - Fix the incorrect behavior of pthread_mutex_init() and std::mutex such that they no longer create a recursive lock by default.
 - `oe_verify_report` now reports the verification result in the `parsed_report` output (only valid when the API returns OE_OK or OE_TCB_LEVEL_INVALID). For the SGX case, the definition of the result is based on `oe_sgx_tcb_status_t`.